requirements.txt
pandas==2.2.3
python_docx==1.1.2
pywebview==5.3.2
pywin32==308
